refactor: move opensearch and elasticsearch classes into packages.
